Generalize MemoizationLeaf Component mixin
Allow any component to easily become a MemoizationLeaf that gets pulled out as
a single unit if any hooks are defined.

Fix REF-1361 and REF-1363
