+++ reflex/components/graphing/recharts/recharts.py
@@ -1,8 +1,7 @@
 """A component that wraps a recharts lib."""
 from typing import Literal
 
-from reflex.components.component import Component, NoSSRComponent
-from reflex.constants import MemoizationDisposition, MemoizationMode
+from reflex.components.component import Component, MemoizationLeaf, NoSSRComponent
 
 
 class Recharts(Component):
@@ -11,35 +10,7 @@ class Recharts(Component):
     library = "recharts@2.8.0"
 
 
-class RechartsMemoizationLeafMixin(Component):
-    """A mixin for Recharts components that must not memoize their children separately.
-
-    This includes all chart types and ResponsiveContainer itself.
-    """
-
-    _memoization_mode = MemoizationMode(recursive=False)
-
-    @classmethod
-    def create(cls, *children, **props) -> Component:
-        """Create a Recharts chart container component (mixin).
-
-        Args:
-            *children: The children components.
-            **props: The props of the component.
-
-        Returns:
-            A Recharts component.
-        """
-        comp = super().create(*children, **props)
-        if comp.get_hooks():
-            # If any of the children depend on state, then this instance needs to memoize.
-            comp._memoization_mode = cls._memoization_mode.copy(
-                update={"disposition": MemoizationDisposition.ALWAYS},
-            )
-        return comp
-
-
-class RechartsCharts(NoSSRComponent, RechartsMemoizationLeafMixin):
+class RechartsCharts(NoSSRComponent, MemoizationLeaf):
     """A component that wraps a recharts lib."""
 
     library = "recharts@2.8.0"
